CRAS: align num_channels format
Also clean unnecessary num_channels parameter.

BUG=None
TEST=unittest

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4648330
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>

+++ cras/src/server/cras_audio_area.c
@@ -11,7 +11,7 @@
 #include "cras/src/server/cras_mix.h"
 #include "cras_audio_format.h"
 
-struct cras_audio_area* cras_audio_area_create(int num_channels) {
+struct cras_audio_area* cras_audio_area_create(size_t num_channels) {
   struct cras_audio_area* area;
   size_t sz;
