+++ cras/src/tests/audio_thread_unittest.cc
@@ -1356,7 +1356,7 @@ int cras_iodev_odev_should_wake(const struct cras_iodev* odev) {
   return 1;
 }
 
-struct cras_audio_area* cras_audio_area_create(int num_channels) {
+struct cras_audio_area* cras_audio_area_create(size_t num_channels) {
   struct cras_audio_area* area;
   size_t sz;
